test-requirements.txt
flask_testing==0.6.1
coverage>=4.0.3
nose>=1.3.7
pluggy>=0.3.1
py>=1.4.31
randomize>=0.13
tox
pymongo  # this should be removed
docker-compose  # this should be also remove
pykube  # this should be also removed